# apiconnect/api.yaml
swagger: '2.0'
info:
  title: Calculator API
  version: 1.0.0
  x-ibm-name: calculator-provider-api
host: $(catalog.host)
schemes:
  - https
basePath: /
produces:
  - application/json
paths:
  /add:
    get:
      summary: Add two numbers
      parameters:
        - name: param1
          in: query
          description: number string
          required: true
          type: string
        - name: param2
          in: query
          description: number string
          required: true
          type: string
      responses:
        '200':
          description: Success
  /sub:
    get:
      summary: Sub two numbers
      parameters:
        - name: param1
          in: query
          description: number string
          required: true
          type: string
        - name: param2
          in: query
          description: number string
          required: true
          type: string
      responses:
        '200':
          description: Success
  /mul:
    get:
      summary: Multiply two numbers
      parameters:
        - name: param1
          in: query
          description: number string
          required: true
          type: string
        - name: param2
          in: query
          description: number string
          required: true
          type: string
      responses:
        '200':
          description: Success
x-ibm-configuration:
  enforced: true
  testable: true
  phase: realized
  cors:
    enabled: true
  assembly:
    execute:
      - operation-switch:
          title: operation-switch
          case:
            - operations:
                - verb: get
                  path: /add
              execute:
                - invoke:
                    title: invoke-add
                    timeout: 60
                    verb: GET
                    cache-response: protocol
                    cache-ttl: 900
                    stop-on-error:
                      - null
                    version: 1.0.0
                    target-url: 'https://sample-apic-app.eu-gb.mybluemix.net/add?param1={param1}&param2={param2}'
            - operations:
                - verb: get
                  path: /sub
              execute:
                - invoke:
                    title: invoke-sub
                    timeout: 60
                    verb: GET
                    cache-response: protocol
                    cache-ttl: 900
                    stop-on-error:
                      - null
                    version: 1.0.0
                    target-url: 'https://sample-apic-app.eu-gb.mybluemix.net/sub?param1={param1}&param2={param2}'
            - operations:
                - verb: get
                  path: /mul
              execute:
                - invoke:
                    title: invoke-mul
                    timeout: 60
                    verb: GET
                    cache-response: protocol
                    cache-ttl: 900
                    stop-on-error:
                      - null
                    version: 1.0.0
                    target-url: 'https://sample-apic-app.eu-gb.mybluemix.net/mul?param1={param1}&param2={param2}'
          otherwise: []
          version: 1.0.0
    catch: []
consumes:
  - application/json